// src/app/core/api/model/dockerHubSearchResultDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DockerHubSearchResultDto { 
    name: string;
    description: string;
    pullCount: number;
    starCount: number;
    isOfficial: boolean;
    isAutomated: boolean;
}