// src/app/core/api/model/imageVerifyResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ImageVerifyResponseDto { 
    exists: boolean;
    digest?: string | null;
    size?: number | null;
    lastUpdated?: string | null;
}